dataset_creation/requirements.txt
hydra-core==1.3.2
jsonlines==4.0.0
numpy==2.3.0
omegaconf==2.3.0
pandas==2.3.0
Requests==2.32.4
tqdm==4.67.1
treelib==1.7.1
